# specs/004-rag-chatbot-openai/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: RAG Chatbot API
  version: 1.0.0
  description: API for the RAG Chatbot integrated with OpenAI Agents and book content.

servers:
  - url: /api
    description: Base API URL

tags:
  - name: Chat
    description: Chatbot interaction endpoints
  - name: Embeddings
    description: Endpoints for managing document embeddings
  - name: Agent
    description: Agent status and utility endpoints
  - name: Health
    description: System health check

paths:
  /chat:
    post:
      summary: Send a message to the RAG Chatbot
      operationId: chatWithBot
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: The user's message.
                session_id:
                  type: string
                  format: uuid
                  description: Optional ID of the existing chat session. If not provided, a new session is started.
                use_agent:
                  type: boolean
                  description: Whether to use the OpenAI Agent for response generation. Defaults to true.
                  default: true
      responses:
        '200':
          description: Successful response from the chatbot.
          content:
            application/json:
              schema:
                type: object
                required:
                  - answer
                  - sources
                  - session_id
                properties:
                  answer:
                    type: string
                    description: The chatbot's generated answer.
                  sources:
                    type: array
                    items:
                      type: string
                    description: Array of citations/sources from the book content.
                  session_id:
                    type: string
                    format: uuid
                    description: The ID of the current chat session.
                  agent_thought_process:
                    type: string
                    nullable: true
                    description: Optional - details about the agent's internal thought process or tool usage.
        '400':
          description: Invalid request payload.
        '500':
          description: Internal server error.

  /chat/selection:
    post:
      summary: Ask the RAG Chatbot about selected text
      operationId: chatWithSelection
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
                - selected_text
              properties:
                message:
                  type: string
                  description: The user's message/question about the selected text.
                selected_text:
                  type: string
                  description: The text selected by the user from the book content.
                session_id:
                  type: string
                  format: uuid
                  nullable: true
                  description: Optional ID of the existing chat session.
      responses:
        '200':
          description: Successful response from the chatbot focusing on selected text.
          content:
            application/json:
              schema:
                type: object
                required:
                  - answer
                  - sources
                  - session_id
                  - focused_on_selection
                properties:
                  answer:
                    type: string
                    description: The chatbot's generated answer.
                  sources:
                    type: array
                    items:
                      type: string
                    description: Array of citations/sources from the book content.
                  session_id:
                    type: string
                    format: uuid
                    description: The ID of the current chat session.
                  focused_on_selection:
                    type: boolean
                    description: Indicates if the answer was primarily focused on the selected text.
        '400':
          description: Invalid request payload.
        '500':
          description: Internal server error.

  /embeddings/generate:
    post:
      summary: Trigger generation/refresh of book content embeddings
      operationId: generateEmbeddings
      tags:
        - Embeddings
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                force_refresh:
                  type: boolean
                  description: If true, regenerate all embeddings even if already present. Defaults to false.
                  default: false
      responses:
        '200':
          description: Embedding generation process initiated or completed.
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - documents_embedded
                properties:
                  status:
                    type: string
                    description: Status of the embedding generation.
                  documents_embedded:
                    type: integer
                    description: Number of documents processed.
        '500':
          description: Internal server error during embedding generation.

  /agent/status:
    get:
      summary: Get status of the OpenAI Agent
      operationId: getAgentStatus
      tags:
        - Agent
      responses:
        '200':
          description: Current status of the agent.
          content:
            application/json:
              schema:
                type: object
                required:
                  - agent_initialized
                  - tools_loaded
                properties:
                  agent_initialized:
                    type: boolean
                    description: True if the OpenAI Agent is initialized.
                  tools_loaded:
                    type: array
                    items:
                      type: string
                    description: List of names of the tools loaded for the agent.
        '500':
          description: Internal server error.

  /health:
    get:
      summary: Check health status of all integrated services
      operationId: getHealthStatus
      tags:
        - Health
      responses:
        '200':
          description: Health status of the application and its dependencies.
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - services
                properties:
                  status:
                    type: string
                    description: Overall health status (e.g., "healthy", "degraded").
                  services:
                    type: object
                    properties:
                      postgres:
                        type: string
                        description: Status of PostgreSQL connection.
                      qdrant:
                        type: string
                        description: Status of Qdrant connection.
                      openai:
                        type: string
                        description: Status of OpenAI API connectivity.
                      agent_sdk:
                        type: string
                        description: Status of OpenAI Agent SDK initialization.